requirements.txt
fastapi
uvicorn
psycopg2-binary
SQLAlchemy
python-jose
python-multipart
passlib[bcrypt]
